+++ yellowstone-grpc-proto/src/plugin/filter/message.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
         geyser::{
-            subscribe_update::UpdateOneof, CommitmentLevel as CommitmentLevelProto,
-            SubscribeUpdate, SubscribeUpdateAccount, SubscribeUpdateAccountInfo,
-            SubscribeUpdateBlock, SubscribeUpdateEntry, SubscribeUpdatePing, SubscribeUpdatePong,
-            SubscribeUpdateSlot, SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
+            subscribe_update::UpdateOneof, SlotStatus as SlotStatusProto, SubscribeUpdate,
+            SubscribeUpdateAccount, SubscribeUpdateAccountInfo, SubscribeUpdateBlock,
+            SubscribeUpdateEntry, SubscribeUpdatePing, SubscribeUpdatePong, SubscribeUpdateSlot,
+            SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
             SubscribeUpdateTransactionStatus,
         },
         plugin::{
@@ -565,14 +565,14 @@ impl DerefMut for FilteredUpdateSlot {
 
 impl prost::Message for FilteredUpdateSlot {
     fn encode_raw(&self, buf: &mut impl BufMut) {
-        let status = CommitmentLevelProto::from(self.status) as i32;
+        let status = SlotStatusProto::from(self.status) as i32;
         if self.slot != 0u64 {
             ::prost::encoding::uint64::encode(1u32, &self.slot, buf);
         }
         if let ::core::option::Option::Some(ref value) = self.parent {
             ::prost::encoding::uint64::encode(2u32, value, buf);
         }
-        if status != CommitmentLevelProto::default() as i32 {
+        if status != SlotStatusProto::default() as i32 {
             ::prost::encoding::int32::encode(3u32, &status, buf);
         }
         if let Some(error) = &self.dead_error {
@@ -581,7 +581,7 @@ impl prost::Message for FilteredUpdateSlot {
     }
 
     fn encoded_len(&self) -> usize {
-        let status = CommitmentLevelProto::from(self.status) as i32;
+        let status = SlotStatusProto::from(self.status) as i32;
 
         (if self.slot != 0u64 {
             ::prost::encoding::uint64::encoded_len(1u32, &self.slot)
@@ -589,7 +589,7 @@ impl prost::Message for FilteredUpdateSlot {
             0
         }) + self.parent.as_ref().map_or(0, |value| {
             ::prost::encoding::uint64::encoded_len(2u32, value)
-        }) + if status != CommitmentLevelProto::default() as i32 {
+        }) + if status != SlotStatusProto::default() as i32 {
             ::prost::encoding::int32::encoded_len(3u32, &status)
         } else {
             0
@@ -986,8 +986,8 @@ pub mod tests {
             plugin::{
                 filter::{name::FilterName, FilterAccountsDataSlice},
                 message::{
-                    CommitmentLevel, MessageAccount, MessageAccountInfo, MessageBlockMeta,
-                    MessageEntry, MessageSlot, MessageTransaction, MessageTransactionInfo,
+                    MessageAccount, MessageAccountInfo, MessageBlockMeta, MessageEntry,
+                    MessageSlot, MessageTransaction, MessageTransactionInfo, SlotStatus,
                 },
             },
         },
@@ -1265,13 +1265,13 @@ pub mod tests {
         for slot in [0, 42] {
             for parent in [None, Some(0), Some(42)] {
                 for status in [
-                    CommitmentLevel::Processed,
-                    CommitmentLevel::Confirmed,
-                    CommitmentLevel::Finalized,
-                    CommitmentLevel::FirstShredReceived,
-                    CommitmentLevel::Completed,
-                    CommitmentLevel::CreatedBank,
-                    CommitmentLevel::Dead,
+                    SlotStatus::Processed,
+                    SlotStatus::Confirmed,
+                    SlotStatus::Finalized,
+                    SlotStatus::FirstShredReceived,
+                    SlotStatus::Completed,
+                    SlotStatus::CreatedBank,
+                    SlotStatus::Dead,
                 ] {
                     encode_decode_cmp(
                         &["123"],
@@ -1289,7 +1289,7 @@ pub mod tests {
                     FilteredUpdateOneof::slot(MessageSlot {
                         slot,
                         parent,
-                        status: CommitmentLevel::Dead,
+                        status: SlotStatus::Dead,
                         dead_error: Some("123".to_owned()),
                         created_at: Timestamp::from(SystemTime::now()),
                     }),